feat: split components in the admin org details
Org Cost centers
